remove unneeded bsv modal component
